requirements.txt
alabaster==0.7.12
appdirs==1.4.3
atomicwrites==1.3.0
attrs==19.3.0
autoflake==1.3.1
Babel==2.7.0
bandit==1.6.2
beautifulsoup4==4.8.2
bidict==0.18.3
black==19.3b0
bleach==3.1.0
bumpversion==0.5.3
certifi==2019.6.16
cftime==1.1.0
chardet==3.0.4
Click==7.0
cocotb==1.2.0
coverage==4.5.4
coveralls==1.8.2
cycler==0.10.0
decorator==4.4.0
doc8==0.8.0
docopt==0.6.2
docutils==0.15.2
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
flake8==3.7.8
flake8-bandit==2.1.1
flake8-broken-line==0.1.1
flake8-bugbear==19.8.0
flake8-builtins==1.4.1
flake8-comprehensions==2.2.0
flake8-debugger==3.1.0
flake8-fixme==1.1.1
flake8-isort==2.7.0
flake8-logging-format==0.6.0
flake8-mutable==1.2.0
flake8-polyfill==1.0.2
flake8-variables-names==0.0.2
funcy==1.14
gitdb2==2.0.5
GitPython==3.0.2
h5py==2.10.0
idna==2.8
imagesize==1.1.0
importlib-metadata==0.20
isort==4.3.21
jdcal==1.4
Jinja2==2.10.1
kiwisolver==1.0.1
lxml==4.5.0
madrigalWeb==3.1.12
MarkupSafe==1.1.1
matplotlib==2.2.2
mccabe==0.6.1
more-itertools==7.2.0
mpmath==1.0.0
netCDF4==1.5.3
networkx==2.4
numpy==1.14.5
openpyxl==2.5.12
packaging==19.1
pandas==0.23.1
parsimonious==0.8.1
pbr==5.4.3
Pebble==4.5.1
pep8-naming==0.8.2
pkginfo==1.5.0.1
pluggy==0.12.0
py==1.8.0
py-aiger==3.9.2
py-aiger-cnf==2.0.0
pycodestyle==2.5.0
pydocstyle==4.0.1
pyeda==0.28.0
pyflakes==2.1.1
PyForecastTools==1.0.1
Pygments==2.4.2
pyparsing==2.2.0
pypblib==0.0.4
pyrsistent==0.15.7
pysatCDF==0.3.0
pytest==5.1.2
pytest-cov==2.7.1
pytest-runner==5.1
python-dateutil==2.7.3
python-dev-tools==2019.9.4
python-sat==0.1.5.dev6
pytz==2018.5
pyupgrade==1.23.0
PyYAML==5.1.2
readme-renderer==24.0
requests==2.22.0
requests-toolbelt==0.9.1
restructuredtext-lint==1.3.0
scikit-learn==0.19.2
scipy==1.3.0
seaborn==0.9.0
six==1.11.0
sklearn==0.0
smmap2==2.0.5
snowballstemmer==1.9.0
soupsieve==2.0
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
stevedore==1.31.0
sympy==1.3
testfixtures==6.10.0
tokenize-rt==3.2.0
toml==0.10.0
toposort==1.5
tox==3.14.0
tqdm==4.35.0
twine==1.13.0
urllib3==1.25.3
virtualenv==16.7.5
wcwidth==0.1.7
webencodings==0.5.1
xarray==0.15.0
zipp==0.6.0
